refactor: change helpers for ConnectedAccountsModal
