+++ src/contentscript/overlay/root/components/SystemPopup/pages/ConnectedAccountsModal/helpers/getSocialOriginTitle.ts
@@ -1,0 +1,9 @@
+import { resources } from '../../../../../../../../common/resources'
+import { IConnectedAccountUser } from '../../../../../../../../common/types'
+
+const getSocialOriginTitle = ([firstAccount, secondAccount]: IConnectedAccountUser[]) =>
+  resources[firstAccount.origin].type === 'social'
+    ? resources[firstAccount.origin].title
+    : resources[secondAccount.origin].title
+
+export default getSocialOriginTitle
